remove more deprecated code (fix some issues)
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1884265 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFCellStyle.java
@@ -182,7 +182,7 @@ public final class HSSFCellStyle implements CellStyle, Duplicatable {
     }
     public void setFont(HSSFFont font) {
         _format.setIndentNotParentFont(true);
-        short fontindex = font.getIndex();
+        short fontindex = (short) font.getIndex();
         _format.setFontIndex(fontindex);
     }
